Initial support for authenticated feeds.
This rather large commit removes the LoginController as it was
only used in one place. It adds an onFeedRequiresAuthentication
method to the FeedLoaderListenerType interface that allows
implementations to block waiting for the user interface to return
credentials via a provided FeedLoaderAuthenticationListenerType
instance. Grouped feed views currently pop up a login dialog when
credentials are required and the cached credentials are not
usable. Ungrouped feed views currently fail to provide any
authentication data and the will likely need further refactoring
of the current interfaces to deal with properly.

References: #117