Force reader HTTP request handling onto a single thread.
This currently causes performance issues when requesting resources
at the beginning of chapters. This may be solvable by caching, as
long as the cache can be emptied at the start of each new book to
prevent serving stale resources from other books that just happen
to have the same names.

References: #137