Do not bother to serve anything for favicon.ico.
Readium seems to request this image so often that it becomes a
bottleneck as there is no caching. The image doesn't appear to be
used anywhere, so getting rid of it entirely is an easy
improvement. The WebViewClient has been moved into a separate
top level class for reasons of clarity.

References: #137