Perform account sync (synchronously) on login.
Perform an account sync on login and don't consider the login to have
succeeded until the sync has either succeeded or failed. This assists
in stopping people from racing to My Books before the contents of their
account has actually been synced. It does not completely prevent it,
but it helps.

Fix #147.