Serve feeds with book database entries
This commit changes the logic of the main FeedLoaderType
implementation: The feed loader now serves feeds that contain
entries from the referenced book database, optionally. This
means that there is no opportunity for the application to
display stale book state, as displayed books always reflect
what is in the local database.

The next step will be to merge in metadata updates when
received by the server: When a book appears in a feed
received from the server, the local book database should
be updated with all non-user-specific information (such as
cover, summary, publisher information, etc).