Add an explicit Revoked state for books
As soon as a book is revoked with the DRM connector, the
book is marked as Revoked. Upon attempting to revoke a book,
a Revoked book is not passed through the DRM connector and
the server is notified of the revocation attempt. If the
server responds with an error, the book remains in the Revoked
state. If the server responds with a success code, the book
is deleted.

Fix #152