Fix calculation of progress within the entire book
The previous approach was based on a heuristic that did not make much
sense and caused the progress bar to jump backwards when moving into a
new spine item. The approach in this commit does not give us a totally
accurate indication of progress but it will do quite well overall and
avoids having to scan through the entire book. The only downside is
the bar will move slightly more per page turn when in shorter spine
items and slightly less per page turn when in longer spine items.