Correct a credentials race
On login success, the existing code immediately schedules a sync
of the account's books. It then saves the credentials that were
received as being correct. The problem here is that the sync task
can run before the credentials have been saved and therefore task
will immediately fail.

The simple fix is to save the credentials first.