Expose a new BookDatabase implementation
This updates the BookDatabase API to expose format information and to
limit the API exposed by each format to operations that make sense for
that format. Applications must now obtain references to "formats" from
database entries and must use the Kotlin "when" expression to obtain
the exact type of each format. This ensures that when new formats are
added, the compiler refuses to compile incomplete "when" expressions,
indicating all of the places in the code that now need to be updated.

The book database implementation has been refactored internally to use
inexpensive re-entrant locks instead of heavyweight file locks. The
file locks were required when the application used separate processes
for the catalog and reader, but this is no longer the case.

Code that appeared to be incorrect in the CatalogReadButton,
CatalogRevokeButton, and CatalogDeleteButton was refactored and/or
replaced.

Testing: All tests pass. New tests were added, although coverage
reporting is inaccurate so the tests are likely incomplete. Verified
that existing books are still noted as being available, and
successfully checked out a non-DRMd book to ensure borrowing still
works.

Affects: http://jira.nypl.org/browse/SIMPLY-1236