Improve BookDatabase implementation and borrowing
This centralizes format information in a place that's not tied to
the BookDatabase implementation (to reduce the number of distinct
places that need to be edited in order to add a new format), and
changes "format" to "format handle" for reasons of taste.

The book borrowing code was converted to Kotlin and refactored so that
as much of the logic as possible is reused for fulfilling non-EPUB
books, and the general assumption that everything is an EPUB should
mostly have been eliminated.

Selection of acquisitions has been refactored and placed into a
central location instead of having views do the work themselves. The
selection criteria for acquisitions has been heavily simplified in
the interests of making it easier for the server to dictate which
formats users should see.

Testing: All tests pass. New tests were added, although coverage
reporting is inaccurate so the tests are likely incomplete. Verified
that existing books are still noted as being available, and
successfully checked out both non-DRMd, DRMd, and "open access but
in an authenticated collection" books to ensure borrowing still works.

Affects: http://jira.nypl.org/browse/SIMPLY-1236