Large commit that fixes revocation and improves testing
The revocation code was suffering from multiple issues, not the
least of which was that it assumed EPUB everywhere and broke
fairly spectacularly when trying to revoke anything else. In the
process of trying to capture bugs in unit tests, it became clear
that the code was nearly impossible to test due to the way that the
controller API had degraded over time. This is still apparent in the
BooksController.bookBorrow() method: It is flat out impossible to
observe the state of the operation in progress - the method operates
asynchronously, does not return any kind of Future-like value, and
does not accept a listener.

This rather large set of changes starts the process of converting
the BookController methods to returning FluentFuture values in
order to make them testable. It also cleans up the internal logic
of some of the tasks in order to make them maintainable, and to
reduce concurrency related issues: Some tasks asynchronously started
other tasks and didn't make any attempt to find out if dependent
tasks had failed or succeeded.

Most importantly: It heavily refactors the book revocation task
and includes a fairly exhaustive set of unit tests to check as
many code paths as is feasible (this includes tests that simulate
DRM failures).

Testing: Borrowed and revoked an audio book successfully.
         Reserved and unreserved an audio book successfully.
         Borrowed and revoked a DRMd EPUB successfully.

Affects: https://jira.nypl.org/browse/SIMPLY-1236