Fix an activity/fragment lifecycle crash
When an activity is destroyed, the current fragment is saved. This
serialized fragment is then passed to the activity's onCreate() method
whenever the system recreates the activity. Unfortunately, this means
that if the fragment is depending on some initialization code having
run in the hosting activity, then the fragment will misbehave: The
fragment's code is run *before* the hosting activity's onCreate()
method body. This can be avoided if the activity simply ignores
whatever saved state was given to its onCreate() method by passing
`null` to the super method.

Testing: All unit tests pass.
         Listening to an audio book and using the developer settings
         to kill activities when the user navigates away no longer
         results in an application crash.