Save and restore player position
This adds the code required to save and restore the player position. It
also explicitly winds the player back to the start of the book if
no saved position is available. This ensures that backends such as
Findaway don't try to erroneously restore the playback position from
other books.

Affects: https://jira.nypl.org/browse/SIMPLY-1323