Do not assume that bookmark lists are mutable
This update strengthens the contract of the database entry interface
to insist that implementations return read-only lists. It adjusts the
bookmark parser to always return (at the very least) unmodifiable
collections. Finally, it fixes the actual bug by creating mutable
copies of the bookmark lists in the code that wants them to be mutable.

Fix: https://jira.nypl.org/browse/SIMPLY-1471