Adjust themes to avoid a crash
With the recent changes to ensure that all activities inherit from
AppCompatActivity, it has been impossible to get as far as the catalog
because the application would crash with:

  "IllegalStateException: You need to use a Theme.AppCompat theme
  (or descendant) with this activity."

This commit deletes the old pre-API 21 themes, removes all references
to themes from the application manifest, and modifies the existing
themes such that they all inherit from Theme.AppCompat. This currently
results in an application with a rather ugly and incorrect color
scheme, but it does at least run.