Do not rewind the book if the user opts to keep it
This reverts part of the change for SIMPLY-391: The book is no longer
told to go back to the first chapter if the user opts not to return
it. It seems that this operation is fairly unreliable with some
audio engines.

Fix: https://jira.nypl.org/browse/SIMPLY-1668
Affects: https://jira.nypl.org/browse/SIMPLY-391