Initial annotation types
This provides a set of Kotlin data classes representing serialized
annotations, based on @gioneill's earlier work. This implementation
uses hand-written Jackson tree-based serialization in preference to
the original reflection-based implementation. The hope is that this
is safer because deserialization is performed on untrusted data
coming from the network.

Affects: https://jira.nypl.org/browse/SIMPLY-1637