Replace last-read and bookmarks API
This implements the logic required for the saving and loading of
the last-read location and the set of per-book user bookmarks. This
removes the code in the ReaderActivity class that manipulates the
old annotation types, and updates the code to use the new APIs. The
code to sync annotations with the remote server has been removed in
preparation for its reintroduction inside the book controller tasks.

The TOC activity has not yet been updated to use the new bookmarks
APIs.

Affects: https://jira.nypl.org/browse/SIMPLY-1637