Ensure that data is loaded for the anonymous profile
This extends the splash screen fragment with a listener method that
is called when the anonymous profile is used. In addition, allow
profileSelect() to be called when the anonymous profile is enabled;
this publishes a selection event so that the controller can load/reload
data for the anonymous profile.

Affects: https://jira.nypl.org/browse/SIMPLY-1637