Replacement bookmark controller and policy API
This implements a new bookmark controller. The main focus of the
design has been on separating the policy (the when and the why
to sync bookmarks) and the mechanism of performing the actual syncing.

To this end, the details of exactly when and why a bookmark will be
synced has been moved into a ReaderBookmarkPolicy type. The implementation
of the policy is expressed in a pure functional form, with the
new ReaderBookController submitting events to the policy for evaluation,
and responding to commands produced by the policy to do the actual work
of sending, receiving, and saving bookmarks.

Previously, bookmark syncing happened for a single account using
management code that was instantiated when the reader was opened, and
discarded some time after the reader was closed. This implementation,
however, starts up a new background thread when the application
starts, and handles bookmark syncing for all accounts in the current
profile. It subscribes to various account and profile events so
internal state is reconfigured correctly when users create/delete
accounts, and switch profiles. It is essentially a perpetually running
service, the lifecycle of which matches that of the application.

Affects: https://jira.nypl.org/browse/SIMPLY-1686