An initial attempt at a no-DRM build
This updates the Gradle scripts to accept a
`org.nypl.simplified.with_drm_adobe` property in a similar manner to
the existing `org.nypl.simplified.with_findaway`.  If the property
is defined in `gradle.properties` and set to `true`, the correct
dependencies for Adobe DRM will be enabled. Additionally, the scripts
now check that you have provided the correct certificate file if you
want DRM, and they check that you have _not_ accidentally provided
it if you are trying to build without DRM (to avoid contaminating
builds with a certificate file accidentally).