Use a linearized acquisition model like iOS
This moves OPDS acquisitions to a model where a set of linearized
acquisition paths are exposed, and the app works in terms of those
paths. This allows the app to make much more intelligent choices
with regards to which acquisition to follow, and should make it
possible to borrow books from Archive.org feeds once the proper
authentication mechanisms are in place. Additionally, the OPDS parser
is no longer responsible for filtering out acquisition entries that
don't provide supported book types - this is now the responsibility
of the higher-level feed loader.

Affects: https://jira.nypl.org/browse/SIMPLY-1788
Affects: https://jira.nypl.org/browse/SIMPLY-1801