Correct bookmark creation
The explicit bookmark button in the reader was not creating bookmarks:
The system was erroneously considering an explicitly-created bookmark
with the same location as an idling bookmark to be equivalent, and
so was ignoring the request to create an explicit bookmark when an
idle bookmark already existed. Because idle bookmarks are automatically
created when viewing a new page, this meant that explicit bookmarks
could *never* be created.

Affects: https://jira.nypl.org/browse/SIMPLY-1637