Remove all traces of the old theme system
This change attempts to locate all places and views in the app that are
still explicitly configuring themselves to use drawables and colors, and
changes them to ensure that they are styled implicitly by the current
theme. This necessitated using a new version of the AudioBooks API that
had been updated to use themes instead of explicit color passing.

Additionally, several bugs were fixed in the reader and audio book
player with respect to using the "current account" instead of the
account that actually owned the book being read/played.

Affects: https://jira.nypl.org/browse/SIMPLY-1691