Fix a serious issue involving crash recovery
When an Android application crashes, the application is not restarted
from a clean slate but instead inherits the stack of activities it
had when it crashed (minus the topmost activity, under the faulty
assumption that the topmost activity must have caused the crash). This
means that the user can bypass the splash screen and fail to configure
the profile system. In builds with the anonymous profile disabled, this
will result in the profile's books not being loaded (so "My Books" will
always be empty). In builds with named profiles, this will result in
hard crashes (resulting in more failed resurrections).

Fixing this involved checking to see if the profile system had ever
been initialized, and forcing the user to go back to the profile
selection screen if it hasn't. The anonymous profile is now selected
unconditionally as soon as the controller is created, which results
in the books being loaded correctly.

In order to test that this works with the profile system, it was
necessary to introduce another app frontend: vanilla-with-profiles.
Correctly getting back to the profile screen when profiles are enabled
required doing some cleanup work in the NavigationDrawerActivity
class due to the initialization of the drawer being dependent on the
profile system.

Affects: https://jira.nypl.org/browse/SIMPLY-1883