Heavily refactor accounts database and credential storage
The accounts database did not expose enough information about the
login status of an account. This made it extremely hard to write
a reasonable user interface that could correctly show login/logout
progress and the current state of any given account. Additionally,
account credentials were stored in external storage, making them
visible to external programs (a security risk).

A new credential storage API was added that, by default, stores
credentials in application-private storage. In order to facilitate
a simple implementation with minimal coupling, the semantics of
the AccountID type were changed such that the AccountID now uniquely
identifies an account across all profiles. The internal representation
is now a UUID.

The accounts database API now exposes credentials via a loginState
field which is a value of an algebraic data type that represents the
exact current state of the account (such as logging in, login failed,
logged in, etc). Credentials are only exposed when an account is in
the logged in state.

The account settings activity has been updated to use the new APIs,
but the login dialog fragment has not.

Affects: https://jira.nypl.org/browse/SIMPLY-1869