Large commit implementing the bulk of the age gate mechanism
The catalog activity was heavily refactored to support an age gate
mechanism. To implement this cleanly, the activity had to be upgraded
to be able to support delaying the loading of a feed (so that an age
gate can be displayed first) and, in the process, it became necessary
to refactor the refresh mechanism too. These changes also fix the
long-broken behaviour regarding the up button: The up button is supposed
to provide hierarchical navigation - the up button moves you "up" the
feeds towards the root of the collection. The back button is supposed
to provide historical navigation _including visiting those feeds that
you visited via the up button_.

In order to implement all of the above sanely, most of the remaining
Feed API types were converted to Kotlin data classes (no more Java
matchers/visitors).

Affects: https://jira.nypl.org/browse/SIMPLY-1636