Remove the RFC3339 module
When the RFC3339 module was first written, Joda Time was unavailable
on Android.  The date and time API provided with Android was a rather
broken implementation from Apache Harmony, and so we imported some
public domain code and used it as the RFC3339 module. As Joda Time
has been stable on Android for years now, and the app has already been
using it for date and time manipulation elsewhere, this change removes
the RFC3339 module and replaces all API uses with Joda Time. This
provides both correctness and performance improvements, and obviously
reduces the amount of code we need to maintain in the app.

Testing: All tests pass
         Casual browsing of the catalog shows no issues

Affects: https://jira.nypl.org/browse/SIMPLY-1907
Affects: https://jira.nypl.org/browse/SIMPLY-915