Add DRM calls
This adds Adobe DRM calls into the login and logout processes. It
adds extensive error reporting, indicating exactly what's going on
in each of the login/logout processes and providing precise details
in the case of failures.

It DOES NOT currently enable DRM for the application: DRM support
is still marked as being unavailable due to a deliberately raised
exception in AdobeDRMServices. The reason for leaving the DRM disabled
at this stage is so that we can check that we've not broken the
application in the non-DRM case.

Testing: Extensive test suite. Tested informally in the NYPL
         collection.

Affects: https://jira.nypl.org/browse/SIMPLY-2010
Affects: https://jira.nypl.org/browse/SIMPLY-1967