Re-enable DRM support!
This re-enables support for Adobe DRM by removing the call that
deliberately raises an exception when trying to instantiate the
connector in AdobeDRMServices.

Currently, borrowing and/or reading encrypted books is not expected
to work.

Testing: Tested logging in and logging out of the NYPL
         collection.

Affects: https://jira.nypl.org/browse/SIMPLY-2010
Affects: https://jira.nypl.org/browse/SIMPLY-1967