Implement borrowing with Adobe DRM
This refactors the borrowing task to enable support for Adobe DRM. It
introduces task step logging, in the same manner as the recent
login/logout work did, for excruciatingly detailed error messages.

Currently, revocation of loans is not yet supported. This will be
the subject of the next commit(s).

This adds extensive testing to try to ensure that every possible
failure case for the DRM connector is handled. The code _may_ currently
be too strict; the DRM connector is known to produce errors that
aren't really errors and that should be ignored. Right now, the code
does not ignore anything.

Testing: Tested borrowing a book from the NYPL collection.

Affects: https://jira.nypl.org/browse/SIMPLY-1638
Affects: https://jira.nypl.org/browse/SIMPLY-1620