Initial pass at account provider descriptions
This introduces a ServiceLoader-based mechanism to populate a registry
with account provider descriptions, and to use them throughout the
app. An account provider description is a minimal description of an
account provider that can be _resolved_ into a full account provider
for the creation of accounts. At the moment, only file-based sources
are supported (equivalent to Accounts.json in the current app).

Currently, the app does not run due to some stubbed out methods
that raise UnimplementedCodeException. This will be addressed in the
following commits.

Affects: https://jira.nypl.org/browse/SIMPLY-1994