Second pass of account provider descriptions
This reimplements enough of the account systems to create and
delete accounts in the application. It changes the account provider
description registry to be a more generalized account provider
registry. Accounts now store full providers on creation, ensuring
that once an account is created, the provider that created it can
never go away.

Affects: https://jira.nypl.org/browse/SIMPLY-1994