Initial server communication
This is a very preliminary and insufficiently-tested version of a
server-based account source.  Currently, this will cause the app to
fail because it does network I/O on the main application thread. The
reason this commit is happening here is because we need to switch to
lazy service initialization on a background thread, and that will
involve fairly intrusive API changes to the main application class
and splash screen.

Affects: https://jira.nypl.org/browse/SIMPLY-1994