Initial refactoring of application startup
The previous commit implemented the initial registry-based account
provider source. Because the list of account providers needed to be
refreshed on application startup, this meant that with the current
application startup scheme, network I/O would have happened on the
main application thread.

This commit refactors the Simplified application class into a "boot"
API that initializes all application services on a background thread,
publishing events as the process proceeds, and publishing an event
on completion (or failure). The boot API exposes a single Future
value that represents the startup in process. The set of application
services are combined into a single SimplifiedServicesType, and the
Future value returns a value of this type if startup succeeds. The
Simplified application class exposes a method that returns a
SimplifiedServicesType directly, for backwards compatibility, and
this method blocks if the application has not yet started up. The
assumption throughout the codebase is that every single activity
other than the splash screen activity can call this method without
blocking in practice.

Additionally, the account provider registry no longer raises
exceptions.

We currently have no sensible user interface for application startup
failures. The exception message is printed and the user can't do
anything other than mash the back button to exit.

Affects: https://jira.nypl.org/browse/SIMPLY-1994