Ensure that we don't miss important boot messages
It's possible for the application to miss boot events when "restarting"
after the last activity has been closed.  By adding a listener to the
boot future, we can guarantee that we won't sit there on the splash
screen waiting for an event that isn't coming.

Affects: https://jira.nypl.org/browse/SIMPLY-1994