Add a disk cache for the NYPL registry
This adds all of the necessary code to support a simple disk cache for
the NYPL registry source. The set of account providers received from
the server are serialized to disk, and on subsequent application runs,
the _first_ time the source is queried, the serialized providers are
returned. Subsequent queries will ask the server for new providers
and update/replace the serialized providers.

Affects: https://jira.nypl.org/browse/SIMPLY-1994