Handle anonymous auth and missing catalog URIs
The specification for the authentication document say that valid
documents _MUST_ contain a catalog URI. In practice, documents aren't
specifying them, so if the document has a missing URI, we take it
from the registry instead.

Additionally, the code was not correctly recognizing anonymous
authentication as a valid authentication type.