Define an initial migration SPI and 3.0 master migration
This defines an SPI that is used to register and execute
migrations. Currently, the only migration defined is from 3.0 master
to the current develop branch.

Because we've never had a release from the current develop branch,
we're also able to pick a sensible set of conventions to ensure easy
future migrations. As such, the application now uses a versioned
base directory:

  val base = context.filesDir
  val vBase = File(base, "4.0")
  val profiles = File(vBase, "profiles")

This ensures that, in future, we do not have to guess which version
of the on-disk data we're migrating _from_ - it is explicit in the
file paths. We also place all Adobe DRM blobs into their own Adobe
subdirectory, in the hope that future migrations will simply involve
copying this directory verbatim instead of having to redo activations.

The current 3.0 master migration code is incomplete; it currently
does the required work to map old-style numeric IDs to a generated
list of UUID-based providers. It extracts credentials and Adobe
device information but stops short of actually creating accounts
at this point.

Because we need strong error reporting for the migration, I've
introduced an initial set of interfaces for the "friendly error
reporting" described in SIMLY-2118, and started retrofitting the
parts of the code that will be called by the migration code.

Additionally, a bug was fixed in the accounts database that would
have allowed creating two accounts with the same provider. This was
documented as being explicitly disallowed but the implementation was
missing the actual check.

Affects: http://jira.nypl.org/browse/SIMPLY-2118
Affects: http://jira.nypl.org/browse/SIMPLY-2068