Harden migrations against failures
This adds additional code to try to ensure that migrations succeed and
to ensure that logging occurs as it should. It also adds the ability
for the app to attach a complete log of everything that was printed
to logcat to the migration report.

Affects: http://jira.nypl.org/browse/SIMPLY-2118
Affects: http://jira.nypl.org/browse/SIMPLY-2068