Ensure that COPPA is not treated as "requiring credentials"
The recent changes to make the COPPA age gate introduced some issues in
that the app now considers an age gate as "requiring credentials". We
now inspect the authentication type more carefully so that we don't
ask for credentials when there won't be any forthcoming.

Affects: http://jira.nypl.org/browse/SIMPLY-2125