Provide more useful event publications
This breaks the profile selection event into two new events:

  ProfileSelectionInProgress
  ProfileSelectionCompleted

The semantics of these events are as follows: The
ProfileSelectionInProgress event is published as soon as the user
selects an accounts, and the task responsible for triggering this
event does all of the must-occur-right-now work such as populating
the book registry. When all of this work is completed, the task
publishes ProfileSelectionCompleted. Consumers should be listening
for ProfileSelectionCompleted unless they have a strong reason to
use the other event.

Affects: http://jira.nypl.org/browse/SIMPLY-1567