Not subbed to BookEvents by default and changes notification intent to SplashActivity.
The behavior for `MainHoldsActivity` was not achieving what we wanted and it looked like the app was bugging out because the top bar said "Reservations" for a bit before reverting to the catalog.

We're experiencing a race condition with OnProfileSelectionInProgress where BookEvents start firing immediately after that ProfileEvent fires, but the NotificationsService doesn't unsubscribe in time. We should start by default with not listening to BookEvents.