Redo the account creation dialog
This reorganizes the account creation dialog to use a fullscreen dialog
and to reflect the fact that the account providers may currently be
in the process of loading. This also adds a "refresh" button to allow
for refreshing the providers from the server.

Fix: https://jira.nypl.org/browse/SIMPLY-2051