Set a document background and foreground aggressively
This sets a foreground and background color in the loaded book
as brutally as possible. For some unknown reason, neither the
ReadiumSDK.reader.setBookStyles function, or the existing
simplified.updateBookStyles function had any effect on the
rendered view. Manually setting the attributes as a pair of
assignments seems to work!

This also adds an override to the web chrome client so that JS
console messages are logged by the host application.

This also fixes a crash that occurred when opening the TOC. The
crash occurred because the layout file was still referring to an old
TabLayout type that had moved to a different package in AndroidX.

Affects: https://jira.nypl.org/browse/SIMPLY-1933