Use the public PDF viewer release
This switches to using the 0.1.0 release of the PDF viewer which was
recently pushed to Maven Central.

Additionally, it adds a command-line property
org.librarysimplified.nexus.publish which, if set to true, attempts
to publish to the NYPL Nexus instead of Maven Central.

Affects: https://jira.nypl.org/browse/SIMPLY-1731
Thanks: @mattniehoff