Add an error activity
This adds a base activity that will be used as the base implementation
until the app is converted to a single-activity setup. This adds a
concrete implementation in the errors package of the app-shared module,
and exports the activity in both vanilla apps. This also adds a button
in the developer settings to start up an error page activity.

Affects: https://jira.nypl.org/browse/SIMPLY-2338