Attach the error page to various parts of the app
This attaches the error page to various operations in the app:

  * Account creation
  * Account deletion
  * Account logins
  * Account logouts
  * Feed loading

Each of these operations, on failure, will provide some sort of link
or button in the UI to open an error page. Various event types have
been augmented to ensure that more contextual information is provided
for display.

Affects: https://jira.nypl.org/browse/SIMPLY-2338