Implement and use a service directory
This change defines a very simple API that application components
can use to obtain instances of dependencies without being strongly
coupled to each other or a singleton. It updates all of the code
that previously referred to the SimplifiedServicesType interface
to instead call methods on a ServiceDirectoryType. It updates the
application boot process to publish services into a directory that
other application components can then consume.

Affects: https://jira.nypl.org/browse/SIMPLY-2313