Replace catalog feed fragment
This replaces the previous catalog feed fragment with a simpler model
such that there is a 1:1 correspondence between feed and fragments,
and a 1:1 correspondence between fragments and their view models. The
reason that this is necessary is because the previous implementation
(using a single view model that maintained a stack of feeds) could
not express a history where a feed leads to a book detail page which
subsequently leads to another feed. This situation can occur with
"Related books" links on book detail pages. Switching to a model
where the fragment backstack is instead used to track history can
easily account for this, but only if one feed == one fragment instance.

Affects: https://jira.nypl.org/browse/SIMPLY-2308
Affects: https://jira.nypl.org/browse/SIMPLY-2287