Continue feeds-without-groups
This continues the implementation of the feeds-without-groups
views. This implements various optimizations to the RecyclerView
configuration (notably the use of a much larger cache for offscreen
views). Additionally, it implements the logic to reconfigure cell views
on book status changes. It refactors the login form to use a ViewModel
to publish dismissal events, although this implementation will likely
be refactored into a shared view model in subsequent commits in order
to eliminate code duplication: The catalog book detail fragment and
the feeds-without-groups views have buttons that need to go through
identical "open a login form, wait until it is closed, then perform
an operation" workflows, so this code could/should be moved into the
view model.

Finally, an activity-scoped "host" view model is introduced in order
to facilitate the injection of services and navigation controllers
into fragments. This allowed for updating the sandbox activity
to better reflect how the actual application will be configured,
and to correctly handle state saving/restoration in the presence
of orientation changes.

Affects: https://jira.nypl.org/browse/SIMPLY-2308
Affects: https://jira.nypl.org/browse/SIMPLY-2287