Change fragment history handling
The current handling of fragment histories had a serious defect: The
BottomNavigator component reused fragments instead of recreating
them, so fragments could persist after a profile change and
would store profile and account IDs that referred to the wrong
profiles/accounts. Essentially, the code was written to assume that
the BottomNavigator would evaluate "get the current profile" lazily
and repeatedly, when in fact it just evaluated it once and stored
the results forever.

The changes made here forcefully delete old fragments using the
BottomNavigator reset() method, and also listen for account deletions
and reset the histories of tabs when accounts are deleted.

Affects: https://jira.nypl.org/browse/SIMPLY-2287