Update sample apps to only increment version code when building on Travis CI.
Previously, the sample apps would increment the version code in
version.properties anytime they were built. This change was not
persisted (committed) so each build would result in the same version
being used (versionCode + 1).

This changes the build script to look for an environment variable named
'TRAVIS_BUILD_NUMBER' and, if present, add that number to the version
from version.properties. This should result in an incrementing version
number for builds run by Travis CI.