Respect input types in authentication documents
This makes the changes necessary to respect the "input" values given
in authentication documents. In practical terms, this means showing
or hiding the username/password fields appropriately both in the
settings account view, and in the popup dialog that appears in the
catalog.

Affects: https://jira.nypl.org/browse/SIMPLY-2674