Save and restore the most recent account
This adds a per-profile preference that used to save and restore the
most recently used account. The identifier of an account is stored
every time the setAccountCurrent() method is called, and the stored
identifier is used when the profile is loaded to set the correct
account.

Affects: https://jira.nypl.org/browse/SIMPLY-2623