Overhaul the login dialog
This adjusts the login dialog so that it is now a full-screen fragment
instead of a Dialog in the Android sense. It hooks up some missing
logic to allow for displaying error details on login failures.

Testing: Tested a successful login on an NYPL account
Testing: Tested a failed login on a Hotchkiss account

Affects: https://jira.nypl.org/browse/SIMPLY-2681