SIMPLY-2694: Retry 403 Forbidden response with auth, when available.
This change fixes a failure with biblioboard where a 403 Forbidden
response is returned instead of a 401 Unauthorized. Instead, of
immediately giving up when a 403 is received, try again with
authentication like we would do with a 401.

Another improvement would be to proactively include the authentication
with the first attempt instead of waiting for a challenge response.