Add a "sync now" analytics event
This adds an analytics event that implementations can respond to by
sending any buffered data to the server. This isn't used by Circulation
analytics (which doesn't buffer anything), but is used by LFA analytics
for debugging. This also adds a button to the developer settings menu
that manually publishes the event.