Add a check to ensure the checksum workaround is in place
This adds a check to the start of the build script
to verify that the build is being performed with the
org.gradle.internal.publish.checksums.insecure property set. This
ensures that the project can be deployed to Central without checksum
issues.

References: https://github.com/gradle/gradle/issues/11308#issuecomment-554317655