Remove obsolete borrow calls that accept accounts
In order to try to track down the cause of SIMPLY-2716, the borrow
calls that take an AccountType parameter have been removed. This
moves the account lookup to the inner borrowing task, allowing for
any exceptions raised to become part of the task failure value.

Affects: https://jira.nypl.org/browse/SIMPLY-2716