Removed simplified-app-vanilla-with-profiles and added new flavor to simplified-app-vanilla.
The simplified-app-vanilla-with-profiles project was largely duplicative
of the simplified-app-vanilla project. With some minor refactoring it
was possible to convert the project to a build flavor of
simplified-app-vanilla. This will make maintaining the sample app much
easier in the future and they will be less likely to drift from one
another.